delet function added

+++ src/Pages/MyList.jsx
@@ -2,6 +2,7 @@ import { useContext, useEffect, useState } from "react";
 import { AuthContext } from "../AuthProvider/AuthProvider";
 import { MdDeleteForever } from "react-icons/md";
 import { FaEdit } from "react-icons/fa";
+import Swal from "sweetalert2";
 
 
 const MyList = () => {
@@ -19,15 +20,34 @@ const MyList = () => {
             });
     }, [user]);
 
-    const handleDelete = (id) =>{
-        console.log(id);
-        fetch(`https://assignment-ten-server-bay.vercel.app/spots/${id}`,{
-            method: 'DELETE'
-        })
-        .then(res => res.json())
-        .then(data =>{
-            console.log(data);
-        })
+    const handleDelete = (_id) => {
+        Swal.fire({
+            title: "Are you sure?",
+            text: "You won't be able to revert this!",
+            icon: "warning",
+            showCancelButton: true,
+            confirmButtonColor: "#3085d6",
+            cancelButtonColor: "#d33",
+            confirmButtonText: "Yes, delete it!"
+        }).then((result) => {
+            if (result.isConfirmed) {
+                fetch(`http://localhost:5000/myProduct/${_id}`, {
+                    method: 'DELETE'
+                })
+                    .then(res => res.json())
+                    .then(data => {
+                        console.log(data);
+                        Swal.fire({
+                            title: "Deleted!",
+                            text: "Your file has been deleted.",
+                            icon: "success"
+                        });
+                        const remaining = mySpots.filter(spo => spo._id !== _id)
+                        setMySpots(remaining)
+                    })
+
+            }
+        });
     }
 
     return (
@@ -57,7 +77,7 @@ const MyList = () => {
                                         <td>{spot.country}</td>
                                         <td>{spot.location}</td>
                                         <td>{spot.cost} taka per person</td>
-                                        <td><button onClick={()=>(handleDelete(spot._id))} className="text-xl text-red-700 "><MdDeleteForever /></button></td>
+                                        <td><button onClick={() => (handleDelete(spot._id))} className="text-xl text-red-700 "><MdDeleteForever /></button></td>
                                         <td><button className="text-xl"><FaEdit /></button></td>
                                     </tr>
                                 )
